requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
python-multipart==0.0.20
pydantic==2.11.7
orjson==3.10.7
openai>=1.30.0
